cmsnet: Added v0.8 version framework in.
The code has a lot of parts requiring completion and revision. Some
parts of the code should be later move into Mininet itself, but for
now they are placed in specific areas inside to contain the necessary
changes to run the CMS network simulator.

A TODO list will be made later.