cmsnet: Added exe/readme on how to run CMSnet
There were previous complaints of how CMSnet was never importing
correctly and always used the /usr/local/... files installed on the VM
instead of the home's mininet folder. These two files help simplify
that. The way to do so is a bit crude (originally desired a more
flexible way that can extend to running Mininet), but certain other
approaches turned out to be a bit tedious.