cmsnet: Moved Mininet patch code to separate file
For clarity, the code to be patched into Mininet later is now in a
separate file. This file will not be run, but the fact that the code is
not commented now makes editing the code easier. This also seemingly
reduces the length of the file (a pointless reason, but some complain).