cmsnet: Bug fixes to running CMSnet
Included a necessary import in CMSnet due to changes to moveLink.
Removed the topo entries in the configurations in case self.topo was
not assigned, in order to prevent saving NoneType as a class name.
Fixed a bug for future mininet patches to node.py with POX paths for
the POXSwitch in the case that the POX_CORE_DIR is not set.