cmsnet.cms_comp: Added config saving protection
In the case that some error occurred while running CMSnet, the config
files for the VMs should not be corrupted. A very specific issue is
on what state the VM should be in (running or not). To make sure that
this state is not corrupted, the config file is not updated until there
is no conflict between the user's view of the state and the actual
state of the VMs.