cmsnet.mininet_node_patch: Use hub controller
The previous POXNormalSwitch uses forwarding.l2_pairs, but this proved
detrimental to cases requiring migrations of VMs. Due to entries not
being able to catch up to these changes (the controller not being in
contact with the CMS and attach()/detach() not called on the switch),
POXNormalSwitch is temporarily switched to using forwarding.hub.