cmsnet.cms_comp: Rewrote the Hypervisor class
Along with format fixing, some of the designs of the Hypervisor class
were also revamped to accommodate future design decisions, including
that of the hypervisor configuration handling.

In the same vein as how the Hypervisor name setter was overridden, the
VirtualMachine class also now does additional tasks to modify that of
the name-to-VM mapping stored in its current hypervisor, if any.