cmsnet.cms_net: Format/revise createVM and related
Fixed formatting of createVM and modified the parameters. Furthermore,
the host class and extra parameters are now actually used in the
function _createHostAtDummy. Changes are also propagated to the patch
file.

Since the header is now quite long due to the number of parameters,
adding any other parameters such as vm_script would have problems. A
possible solution is to just make "host_cls" as "cls" and squash it
into params, but this does have problems with clarity. On the other
hand, those that are familiar with Mininet may actually be used to
having "cls" be for the node's class. The decision will be left to when
the vm_script has to be added in.