cmsnet.cms_net: Separate default HV name handling
Code for handling default hypervisor names is now extracted and placed
in a different method. This is in hope that the method can be used for
other methods later such as invict. Furthermore (and more importantly),
doing so makes launchVM cleaner and this portion itself may require a
bit of rewriting/format fixing later.