cmsnet.cms_net: Rewrote HV name handling
Method _getNextDefaultHVName is rewritten to accommodate more cases
and to make the code more Pythonic. A previous logic problem with
the packed mode was fixed in the process.

Unfortunately, since there was no clean and readable way to find the
maximum amount with a given limiting/filtering function (and a possible
default value), the current packed mode handler loops through twice,
once to filter and another to find the current hypervisor. There is a
more efficient version based on the original code by guoluyi, but it is
less compact. The two should do more or less the same thing.