cmsnet.cms_net: Rewrote HV name handling again
Method _getNextDefaultHVName is rewritten again, now with a more
flexible way of handling the distribution of VMs. The base function
is implemented similar to POX message handling, so readability may
decrease from this. On the other hand, this allows for being able to
easily add new modes like "same" or "next" to the list, although these
new modes may not be useful as of the moment.

Some changes were also made to the __init__ method to allow finding
of the possible modes dynamically. A thought is of whether the newly
written method getPossibleVMDistModes should be a class method instead,
so that the cms executable can also call it without instantiating.