cmsnet: Rewrote/format cloneVM and related code
There were some issues with the previous merged code that did not make
correct configurations on the intended VM. Some assertions also don't
seem to pass if given a bad input. Finally, there were many problems
with naming conventions and inputs, currently not up-to-date with all
other pieces of code.

Some changes hint at possible changes to the design of other parts of
the code, which may show up in a later commit.