cmsnet.cms_cli: Update CLI code aliasing
To help the usage error message be more clear, the command name can be
passed as a value to ease aliasing of commands. VM name checking now
does not check running status by default, and it skips names that are
not set. Finally, vm_script passing is now implemented, which involves
code in the CMSnet class as well.