cmsnet.cms_cli: Redesigned input checking methods
The checking methods were originally designed to be too bulky and had
combined all possible conditions together, reducing the readability of
the various tests. The redesigning is in hopes that there is a sort of
hierarchy to the checks involved, considering subcases within cases.
Furthermore, this allows for future designs to possibly use component
instances as input instead of just their names.

A noteworthy beta design made in this stage is for checks to also
consider the pausing functionality that may be used in later designs.