cmsnet.cms_net: Revise mode changing to use old_hv
Modified all relevant methods dealing with vm_dist_mode to consider
the possibility of the VM's previous hypervisor, so that even migrateVM
can get a default hypervisor if necessary (easing the work for evicting
the VMs later). Modes next and prev are now turned into cycleall, so
as to remove dependency on the last_hv attribute.

Configuration writing will be done at a later point, as considering how
the attributes can be saved is another issue being resolved.