cmsnet.cms_comp: Revised config handling again
The error printing was refactored, each part seemed to take up too much
space and was a completely different part of the logic. Furthermore,
tracebacks are also printed since errors occurring at this part already
signifies that something when severely wrong and require debugging.

The code in the configuration handling was also simplified for sake of
readability, combining try/except blocks that didn't need to be split.