cmsnet.cms_net: Revise CMSnet config handling
Propagate the changes from the CMS components to the CMSnet class as
well, including changes to the structure of the configuration handling
methods, the use of config_error, and finally the config file locking.
Also, the use of the individual components' config locking is also done
when setting up, allowing the ability to preserve the components' config
files while trying to get back the old status.