cmsnet: Implement VM pausing and resuming
Complete implementation for pausing and resuming VMs. The CLI also now
uses the beta checking method as default, since pausing is now a valid
action to take. Currently, the underlying implementation only uses a
flag and two scripts in the VM class. As the number of scripts are
to increase, there may be a need to group them all together as one
dictionary, mapping from script type to script file locations.

A thought while implementing this part was on whether hypervisors should
also have it disabled upon setup rather than directly modifying the flag
attribute, but this seems unnecessary as of now because enabling only
requires this flag and eviction of VMs (of which is unnecessary upon
starting CMSnet because no hypervisors will have any VMs by default).