cmsnet: Move Mininet patch methods to MininetPatch
Because needing to keep track of two different versions was a hassle
while developing code, it was simpler to just have a Mininet subclass
called MininetPatch that happens to have all the methods and changes
it needs for CMSnet to work. This ways, CMSnet will need the minimal
amount of changes to adjust to the new additions once they are actually
in Mininet.