cmsnet.cms_api: Implemented DH API (excp vers.)
Implemented the debatably hacky API with raising exceptions as a basis
for reacting against errors instead of passing booleans. The mode
changing is currently semi-incomplete due to the complexity of the
other modes, but the current basic 3 are more than sufficient to test
out the CMS network. The exception classes are also somewhat
disorganized and may need to be revised or rewritten later.