cmsnet: Update CMSnet with changes in Mininet
Update CMSnet with the corresponding changes in Mininet applied, so as
to be up-to-date with Mininet functionalities. This only applies to
the slight changes in the CLI and the Mininet/CMSnet class.

Mininet is currently not merged in at this point, since duplicate
commits cause rebasing and patching to be difficult. This commit is only
intended to keep CMSnet up-to-date with Mininet changes so far.