cmsnet.cms_net: Implement getPossibleVMScripts
Implement the method to get VM script names dynamically. Since the
scripts are neither generated by a Python module nor automatically in
the configuration folder, there is a need to search for it among the
system paths.

The problem is that VMs themselves would have to know the scripts
folder's path as well. This can be solved either by passing the path
information too or by copying the necessary files to the configurations
folder (which VMs currently do know about). Both ways have problems,
the former being too verbose and the latter being harder to clean up.
Both also don't allow changing of the scripts later.