cmsnet.cms_net: Collect CMSnet info needed by comp
Put together a dictionary of information about the CMSnet class that
may be needed by the individual components. This includes the previously
passed config_folder attribute. By passing the dictionary, not only does
this allow more information to be passed about the CMSnet, but it also
maintains consistency if CMSnet decides to change these values. Some
of the complexity can even be pushed to the components instead, and this
also allows components to change attributes while keeping consistency.