cmsnet.cms_comp: Implement VM script handling
Completely implemented the handling of the VM script running. Like how
the script files are organized, the component only needs to know the
folder location to run each of the type of scripts inside.

There is a bit of trouble at implementing the parameters passing for the
scripts, as a dictionary does not activate the setter method if modified
internally. Hence, it was necessary to create a new subclass that would
update the VM's configurations when necessary. For organization sake, it
may be moved to a different file later. Also, keeping a specific order
of all arguments is difficult, so it may be necessary to have them in a
key-value format. This is unnecessary for the current scripts, though.