cmsmsg.cmsnet_service: Move info classes
Moved HVInfo and VMInfo into the cmsnet_service code to simplify the
controller code. This is so that the controller itself won't have to
do any parsing of the data sent with the CMS message.

Some other code was also simplified and/or refactored because of this
change, and the class hierarchy of events is more flattened since the
code is too simple to require refactoring.