cmsmsg.knowing_beta: Revise/rewrote handle methods
Revise the _handle_ methods based on the new attributes of the events
raised by CMSBot. Most of this just has to do with changing how to
access necessary information to send to switches. Connections are now
extracted directly from the POX core instead of being saved locally.

A noteworthy concern is that while testing out functionalities of POX
in extracting Connection instances and port numbers, it seems that there
is a bug in either the PortCollection class or related code , where
previous instances of the port remain, sharing names but with different
port numbers. Currently there is a workaround (as a separate method),
although it may be better to have this fixed later on.