cmsnet.cms_comp: Implement script rc file writing
Implement the writing of the rc file for the VM script, a workaround
for the previous problem of argument order. Furthermore, the home
directory is now set to the temporary folder. There is also an attempt
to kill the start script process, although this hopefully is cleaned
up already by the user.