cmsnet.vm_scripts: Revise scripts to first check
Added a check script to first check the environment variables. All of
the checks for the environment variables should go here, so that before
any scripts are run, the check script must be ran first. It is fine if
there is no check if there is no point in checking.