cmsnet.cms_util: Redesign util module on addresses
Redesigned the parsing and validating functions on address handling, and
added more methods to expand partial IP and MAC addresses if used (i.e.
having an IP of 10.6 is equivalent to 10.0.0.6). Validation of the IP
now uses the socket module, although there isn't a good equivalent for
MAC addresses. Also, ipParse is redefined since the Mininet version is
restrictive. Finally, the VM class methods were also revised to adjust
to these changes.