cmsnet: Remove interface dependency on node name
Changed all [nodename]-X interface names to host-X, as almost all hosts
that will ever be used will have their own network namespace, preempting
any conflicts in different node's interface names. In the case that the
host does not have its own network namespace, the interface name is
defaulted to the original [nodename]-X format, with the 15 character
limit as noted before. Also, this depends on the notion that VMs will
never be created in parallel, so any attempts to do so will require some
more changes to the code.

As a side effect, hypervisors are now forbidden to use the name "host",
even if they have their own preset DPID.