cmsnet: Fix CMSnet.stop() to guard against errors
Fixed the stop() method and all related methods to not abort when an
error is encountered, so that cleanup will not stop midway through.

The API command line is also fixed to actively stop, so that CMSnet
will not stall after finishing execution due to the separate thread
that tries to keep alive the connection to the controller.