cmsnet.cms_comp: Add alias to netmask and gateway
Added alias attributes to netmask and default_gateway (subnet_mask and
gateway respectively). Also switched the importance of gateway and
default_gateway, as the shorter one is likely to be used more. In the
case that it later becomes unclear unless "default" is specified, the
info default dictionary and such will be changed back.