cmsnet: Readjusted terminal setup
Readjusted terminal setup for both xterm and gterm to use node.cmd() to
start up the terminals instead of the Popen class in the subprocess
module. The advantage to doing so is that using node.cmd() maintains the
environment variables and folder location, so that CMSnet can set up a
different home folder for each node. Unfortunately, once the terminal
starts up, any changes made through the terminal does not persist.

The method CMSnet.makeTerms should be later changed to makeTerm due to
only processing one component at a time.