cmsnet.cms_net: Use self to invoke classmethod
Replace CMSnet.method() with self.method(), despite the method being a
class method instead of that of an instance. In a similar logic to the
mininet_link_patch, using self would allow the flexibility for CMSnet's
subclasses to override the class method's functionality. Since CMSnet
is already initialized at this point, there is no harm in using it this
ways.