cmsnet: Add checks on node name lengths
Added various checks and logging for cases that the user suggests a node
name longer than 13 characters. This is due to the interface name in
Linux being restricted to 14 characters, of which the last 2 are for
differentiating between interfaces. This only applies to nodes in the
public namespace like switches, so hosts in their own private network
namespace do not need to adhere to this rule.